# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Role": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"menus": {
						"items": {
							"$ref": "#/components/schemas/MenuRole"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"menus",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isActive": {
						"type": "boolean"
					},
					"auth": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					},
					"SAPCode": {
						"type": "string"
					},
					"SAPUser": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"username",
					"password",
					"email",
					"role",
					"isActive",
					"auth",
					"created",
					"operation",
					"area",
					"line",
					"SAPCode",
					"SAPUser"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUserPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isActive": {
						"type": "boolean"
					},
					"SAPCode": {
						"type": "string"
					},
					"SAPUser": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"role",
					"isActive",
					"SAPCode",
					"SAPUser",
					"operation",
					"area",
					"line"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Responsable": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operationId": {
						"$ref": "#/components/schemas/Operation"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"operationId",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Cause": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"groupCode",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Symptom": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"groupCode",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Notice": {
				"properties": {
					"id": {
						"type": "string"
					},
					"numNotice": {
						"type": "number",
						"format": "double"
					},
					"otCode": {
						"type": "string"
					},
					"didCard": {
						"type": "string"
					},
					"failureTime": {
						"type": "string"
					},
					"departmentId": {
						"$ref": "#/components/schemas/Area"
					},
					"lineId": {
						"$ref": "#/components/schemas/Line"
					},
					"cardTypeId": {
						"$ref": "#/components/schemas/Card"
					},
					"cardTitle": {
						"type": "string"
					},
					"priorityId": {
						"$ref": "#/components/schemas/Priority"
					},
					"componentsId": {
						"$ref": "#/components/schemas/Component"
					},
					"breakdownId": {
						"$ref": "#/components/schemas/Breakdown"
					},
					"failureTypeId": {
						"$ref": "#/components/schemas/TypeFail"
					},
					"affectsId": {
						"$ref": "#/components/schemas/Affect"
					},
					"affectsFile": {
						"type": "string"
					},
					"cardDescription": {
						"type": "string"
					},
					"processId": {
						"$ref": "#/components/schemas/Process"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"sapId": {
						"type": "string"
					},
					"equipmentId": {
						"$ref": "#/components/schemas/LineMachine"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"objectId": {
						"$ref": "#/components/schemas/ObjectParts"
					},
					"responsableId": {
						"$ref": "#/components/schemas/Responsable"
					},
					"causeId": {
						"$ref": "#/components/schemas/Cause"
					},
					"symptomId": {
						"$ref": "#/components/schemas/Symptom"
					},
					"textCause": {
						"type": "string"
					},
					"textSymptom": {
						"type": "string"
					},
					"urlPhoto": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"otCode",
					"didCard",
					"failureTime",
					"departmentId",
					"lineId",
					"cardTypeId",
					"cardTitle",
					"priorityId",
					"componentsId",
					"breakdownId",
					"failureTypeId",
					"affectsId",
					"affectsFile",
					"cardDescription",
					"processId",
					"user",
					"sapId",
					"equipmentId",
					"operation",
					"objectId",
					"responsableId",
					"causeId",
					"symptomId",
					"textCause",
					"textSymptom",
					"urlPhoto",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INoticePayload": {
				"properties": {
					"otCode": {
						"type": "string"
					},
					"didCard": {
						"type": "string"
					},
					"failureTime": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					},
					"cardType": {
						"$ref": "#/components/schemas/Card"
					},
					"cardTitle": {
						"type": "string"
					},
					"priority": {
						"$ref": "#/components/schemas/Priority"
					},
					"components": {
						"$ref": "#/components/schemas/Component"
					},
					"breakdown": {
						"$ref": "#/components/schemas/Breakdown"
					},
					"failureType": {
						"$ref": "#/components/schemas/TypeFail"
					},
					"affects": {
						"$ref": "#/components/schemas/Affect"
					},
					"affectsFile": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"cardDescription": {
						"type": "string"
					},
					"process": {
						"$ref": "#/components/schemas/Process"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"sapId": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"objectId": {
						"$ref": "#/components/schemas/ObjectParts"
					},
					"causeId": {
						"$ref": "#/components/schemas/Cause"
					},
					"symptomId": {
						"$ref": "#/components/schemas/Symptom"
					},
					"textCause": {
						"type": "string"
					},
					"textSymptom": {
						"type": "string"
					}
				},
				"required": [
					"otCode",
					"didCard",
					"failureTime",
					"department",
					"line",
					"cardType",
					"cardTitle",
					"priority",
					"components",
					"breakdown",
					"failureType",
					"affects",
					"affectsFile",
					"cardDescription",
					"process",
					"user",
					"sapId",
					"operation",
					"objectId",
					"causeId",
					"symptomId",
					"textCause",
					"textSymptom"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INoticenPayloadNewFormat": {
				"properties": {
					"processId": {
						"$ref": "#/components/schemas/Process"
					},
					"didCard": {
						"type": "string"
					},
					"failureTime": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"lineId": {
						"$ref": "#/components/schemas/Line"
					},
					"equipment": {
						"$ref": "#/components/schemas/Equipment"
					},
					"cardTypeId": {
						"$ref": "#/components/schemas/Card"
					},
					"cardTitle": {
						"type": "string"
					},
					"priorityId": {
						"$ref": "#/components/schemas/Priority"
					},
					"componentsId": {
						"$ref": "#/components/schemas/Component"
					},
					"breakdownId": {
						"$ref": "#/components/schemas/Breakdown"
					},
					"failureTypeId": {
						"$ref": "#/components/schemas/TypeFail"
					},
					"cardDescription": {
						"type": "string"
					},
					"affectsId": {
						"$ref": "#/components/schemas/Affect"
					},
					"otCode": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"affectsFile": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"isActive": {
						"type": "boolean",
						"enum": [
							"true"
						],
						"nullable": false
					},
					"sapId": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"objectId": {
						"$ref": "#/components/schemas/ObjectParts"
					},
					"causeId": {
						"$ref": "#/components/schemas/Cause"
					},
					"symptomId": {
						"$ref": "#/components/schemas/Symptom"
					},
					"textCause": {
						"type": "string"
					},
					"textSymptom": {
						"type": "string"
					}
				},
				"required": [
					"processId",
					"didCard",
					"failureTime",
					"department",
					"lineId",
					"equipment",
					"cardTypeId",
					"cardTitle",
					"priorityId",
					"componentsId",
					"breakdownId",
					"failureTypeId",
					"cardDescription",
					"affectsId",
					"otCode",
					"id",
					"affectsFile",
					"userId",
					"isActive",
					"sapId",
					"operation",
					"textCause",
					"textSymptom"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Field": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fieldType": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"optionsEndpoint": {
						"type": "string"
					},
					"validations": {
						"type": "string"
					},
					"childfield": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"fieldType",
					"label",
					"optionsEndpoint",
					"validations",
					"childfield"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFieldPayload": {
				"properties": {
					"fieldType": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"optionsEndpoint": {
						"type": "string"
					},
					"validations": {
						"type": "string"
					},
					"childfield": {
						"type": "string"
					}
				},
				"required": [
					"fieldType",
					"label",
					"optionsEndpoint",
					"validations",
					"childfield"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Consecutive": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"operation",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IConsecutivePayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"SAPCode",
					"operation",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ICausePayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"SAPCode",
					"groupCode",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISymptomPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"SAPCode",
					"groupCode",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponsablePayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					}
				},
				"required": [
					"name",
					"SAPCode",
					"operation"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fixmaint",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Richard HC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "take",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"put": {
				"operationId": "UpdateUserStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/notices": {
			"get": {
				"operationId": "GetNotices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Notice"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "top",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateFrom",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateEnd",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sapForm",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isWeb",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeFrom",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeEnd",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "operationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "totalRows",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNoticeNewFormat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Notice"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INoticenPayloadNewFormat"
							}
						}
					}
				}
			}
		},
		"/notices/old_notice": {
			"post": {
				"operationId": "CreateNotice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Notice"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INoticePayload"
							}
						}
					}
				}
			}
		},
		"/notices/{id}": {
			"get": {
				"operationId": "GetNotice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Notice"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/fields": {
			"get": {
				"operationId": "GetFields",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Field"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Field"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateField",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Field"
								}
							}
						}
					}
				},
				"tags": [
					"Field"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IFieldPayload"
							}
						}
					}
				}
			}
		},
		"/fields/{id}": {
			"get": {
				"operationId": "GetField",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Field"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Field"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/consecutives": {
			"get": {
				"operationId": "GetConsecutives",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Consecutive"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Consecutive"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateConsecutive",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Consecutive"
								}
							}
						}
					}
				},
				"tags": [
					"Consecutive"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConsecutivePayload"
							}
						}
					}
				}
			}
		},
		"/consecutives/{id}": {
			"get": {
				"operationId": "GetConsecutive",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Consecutive"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consecutive"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/causes": {
			"get": {
				"operationId": "GetCauses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Cause"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cause"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "groupCode",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCause",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Cause"
								}
							}
						}
					}
				},
				"tags": [
					"Cause"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICausePayload"
							}
						}
					}
				}
			}
		},
		"/causes/{id}": {
			"get": {
				"operationId": "GetCause",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Cause"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cause"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/symptoms": {
			"get": {
				"operationId": "GetSymptoms",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Symptom"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Symptom"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "groupCode",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSymptom",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Symptom"
								}
							}
						}
					}
				},
				"tags": [
					"Symptom"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISymptomPayload"
							}
						}
					}
				}
			}
		},
		"/symptoms/{id}": {
			"get": {
				"operationId": "GetSymptom",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Symptom"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Symptom"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/responsables": {
			"get": {
				"operationId": "GetResponsables",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Responsable"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Responsable"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "operation",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateResponsable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Responsable"
								}
							}
						}
					}
				},
				"tags": [
					"Responsable"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IResponsablePayload"
							}
						}
					}
				}
			}
		},
		"/responsables/{id}": {
			"get": {
				"operationId": "GetResponsable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Responsable"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Responsable"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}